make the login redirect work with passing the right props
