+++ client/src/components/profile/DisplayVolunteerProfile.tsx
@@ -1,39 +1,44 @@
-import React from "react";
+import React, {useEffect, Props} from "react";
 import VolunteerProfile from "./VolunteerProfile";
 import { gql } from "apollo-boost";
 import { useQuery } from "@apollo/react-hooks";
 import { useSelector } from "react-redux";
 
-const GET_VOLUNTEER_PROFILE = gql`
-  query GetVolunteerProfile($contact_email: String!) {
-    user(contact_email: $contact_email) {
-      first_name
-      last_name
-      contact_email
-      contact_phone
-      description
-    }
-  }
-`;
-export default function DisplayVolunteerProfile() {
+// const GET_VOLUNTEER_PROFILE = gql`
+//   query GetVolunteerProfile($contact_email: String!) {
+//     user(contact_email: $contact_email) {
+//       first_name
+//       last_name
+//       contact_email
+//       contact_phone
+//       description
+//     }
+//   }
+// `;
+export default function DisplayVolunteerProfile(props: any) {
   const contact_email = useSelector(
     (state: any) => state.app.userInfo && state.app.userInfo.email
-  );
+  )
+  
+  
+  // const { loading, error, data } = useQuery(GET_VOLUNTEER_PROFILE, {
+  //   variables: { contact_email }, 
+  //   pollInterval: 300,
+  // });
 
-  const { loading, error, data } = useQuery(GET_VOLUNTEER_PROFILE, {
-    variables: { contact_email }
-  });
+  // if (loading) {
+  //   return <p>Loading...</p>;
+  // }
 
-  if (loading) {
-    return <p>Loading...</p>;
-  }
-
-  if (error) {
-    return <p>{error.message}</p>;
-  }
+  // if (error) {
+  //   return <p>{error.message}</p>;
+  // }
+  
+  // console.log('data: ', data);
+  
   return (
     <div className="container">
-      <VolunteerProfile {...data.user} />
+      {contact_email && <VolunteerProfile contact_email={contact_email} />}
     </div>
   );
 }
